feat(controller) : separate the controller function of modifing data and judge the flow
Add controlling module about information of programs :
- Game : 모든 게임 진행
- InputReceiver : Input을 받고 유효성 판단 및 저장
- RandomBallSet : 랜덤으로 볼세트를 생성하는 객체
- RuleChecker : 유저의 점수를 채점하는 객체
