+++ src/main/java/baseball/controller/RandomBallSet.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.controller;
 
 import utils.RandomUtils;
